Disconenct > LeaveRoom
Now people are disconnecting entirely when they go back to main menu and connect again in 3 seconds.

Previously they were leaving the room and reconnecting to the master server with LeaveRoom() function. But, this was too quick and increasing the CCU artificially for 5 seconds. Now, the CCU is actual CCU.