DB Snapshot
Now whenever someone updates the top users list, we take a snapshot of all users of that week.

Since we are already reading all users of that week, its is nothing to compress them in a few documents and write to the DB.

If we have 20,000 weekly players. Every update will read 20,000 docs but will write only 10 docs.

+ Bug fix: We were not updating the last login xD